[fix#Subscription] 회원의 구독한 강사 리스트 조회 API 수정 S13P11E104-260 1. InstructorQueryService에서 queryUserSubscriptionsByUserId 해당 강사의 구독자 수 조회 코드 수정 2. subscriptionQueryService.countSubscribers 반환타입 Integer로 변환 3. InstructorQueryService에서 queryInstructorProfile 해당 강사 구독자수 조회 코드 수정

+++ Backend/src/main/java/com/e104/reciplay/user/subscription/service/SubscriptionQueryServiceImpl.java
@@ -20,7 +20,7 @@ public class SubscriptionQueryServiceImpl implements SubscriptionQueryService {
     }
 
     @Override
-    public Long countSubscribers(Long instructorId) {
+    public Integer countSubscribers(Long instructorId) {
         return subscriptionRepository.countByInstructorId(instructorId);
     }
